requirements.txt
africastalking==1.2.3
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
Django==3.1.2
django-dynamic-fixture==3.1.0
django-dynamic-fixtures==0.2.1
django-phonenumber-field==5.0.0
flake8==3.8.4
idna==2.10
install==1.3.4
mccabe==0.6.1
mock==4.0.2
mypy-extensions==0.4.3
pathspec==0.8.0
phonenumbers==8.12.11
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.3
pytz==2020.1
regex==2020.10.11
requests==2.24.0
schema==0.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
